# Improve performance of get_objects_for_user